+++ Module.php
@@ -8,10 +8,13 @@
 
 namespace humhub\modules\cleanTheme;
 
+use humhub\libs\DynamicConfig;
+use humhub\modules\ui\view\helpers\ThemeHelper;
 use Yii;
 
 class Module extends \humhub\components\Module
 {
+    const BASE_THEME_NAME = 'clean-base';
 
     /**
      * @var string defines the icon
@@ -37,5 +40,59 @@ class Module extends \humhub\components\Module
     public function getDescription()
     {
         return Yii::t('CleanThemeModule.config', 'Clean theme for Humhub based on the Community theme');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function disable()
+    {
+        $this->disableTheme();
+        parent::disable();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function enable()
+    {
+        if (parent::enable()) {
+            $this->enableTheme();
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * @return void
+     */
+    private function enableTheme()
+    {
+        // Already a theme based theme is active
+        foreach (ThemeHelper::getThemeTree(Yii::$app->view->theme) as $theme) {
+            if ($theme->name === self::BASE_THEME_NAME) {
+                return;
+            }
+        }
+
+        $theme = ThemeHelper::getThemeByName(self::BASE_THEME_NAME);
+        if ($theme !== null) {
+            $theme->activate();
+            DynamicConfig::rewrite();
+        }
+    }
+
+    /**
+     * @return void
+     */
+    private function disableTheme()
+    {
+        foreach (ThemeHelper::getThemeTree(Yii::$app->view->theme) as $theme) {
+            if ($theme->name === self::BASE_THEME_NAME) {
+                $ceTheme = ThemeHelper::getThemeByName('HumHub');
+                $ceTheme->activate();
+                break;
+            }
+        }
     }
 }